Fix grep command in mountcd.sh for Dell nodes
When using an ISO URL that contains an IPv6 address including brackets,
those brackets are interpreted by grep as part of a regexp, causing
the mountcd.sh script to fail.

If we run "grep -F", the ISO string will be taken as a literal, which is
more in line with the expected outcome.